refactor: simplify e2e test folder structure
