+++ tests/e2e/cucumber/features/smoke/searchProjectSpace.ocis.feature
